refactor around action order
